refactor implementation-neutral inner class
